slack-ai-assistant/requirements.txt
slack_bolt
openai
notion-client
schedule
python-dotenv
